refactor(layout): move domain logic into domain/
Step 3 of the domain/ reorganization plan. Moves the three modules
that own LCM's runtime behavior into domain/:

* coordinator.py -> domain/coordinator.py
  LockUsercodeUpdateCoordinator -- per-lock polling/push
  coordinator with backoff.

* sync.py -> domain/sync.py
  SlotSyncManager and SlotState -- the tick, state machine, and
  per-slot circuit breaker.

* slot_manager.py -> domain/slot_coordinator.py
  SlotEntityCoordinator (the per-slot intent-dispatch surface) and
  PinRequiredError. Renamed to match the class inside; the old
  name suggested CRUD slot management.

tests/test_slot_manager.py -> tests/test_slot_coordinator.py to
track the canonical name.

Internal relative imports within the moved files shifted one
level up for non-domain siblings (.const -> ..const, .models ->
..models, .util -> ..util, .providers -> ..providers under
TYPE_CHECKING) and down for new siblings inside domain/
(.domain.exceptions -> .exceptions, .domain.queries -> .queries,
.domain.config -> .config, .domain.resilience -> .resilience).

Also fixed a pre-existing latent bug in domain/exceptions.py:
its TYPE_CHECKING ``from .providers import BaseLock`` was wrong
relative-path-wise after the Step 1 move (would have looked for
domain/providers/) but never crashed because it ran only under
type-checking, not at runtime. Now corrected to ..providers.

External importers and string-based test mocks
(custom_components.lock_code_manager.coordinator.X etc.) updated
to the new paths.
